html/testdata/go/issues.dat

#data
<html><head></head><body><tag1><tag2 /><p></p></tag1><div></div></body></html>
#errors
#document
| <html>
|   <head>
|   <body>
|     <tag1>
|       <tag2>
|         <p>
|     <div>

#data
<table><math><th><mo><select></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math th>
|         <math mo>
|           <select>
|     <table>
